% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPI.R
\name{circ_protein_interaction}
\alias{circ_protein_interaction}
\title{Prediction of interaction between circRNA and protein}
\usage{
circ_protein_interaction(circRNA_seq, protein_seq, output_folder)
}
\arguments{
\item{circRNA_seq}{A dataframe containing the names and sequences of circRNAs}

\item{protein_seq}{A dataframe containing the names and sequences of proteins}

\item{output_folder}{The name of the output folder}
}
\value{
The interaction between the circRNA and protein will written in
output_folder
}
\description{
This function takes the circRNA and protein sequences as input and predicts
the interaction between them.
}
\examples{
#Loading an example data for circRNA names and sequences.
a<-data("circRNA_seq_df")
circ_df<-circRNA_seq_df

#Loading an example data for protein names and sequences.
b<-data("protein_seq_df")
protein_df<-protein_seq_df

#predicting circRNA and protein interaction. Here, the output will be
#written in file interaction_data.txt in out_dir directory
out_dir<-tempdir()
circ_protein_interaction(circ_df, protein_df, out_dir)



}
