\name{circRNA_seq_df}
\alias{circRNA_seq_df}
\docType{data}
\title{
  A sample dataframe of circRNAs names and sequences.
}
\description{
  This data is a sample data containing the names and sequences of circRNAs.
}
\usage{data("circRNA_seq_df")}
\format{
  A data frame with 5 observations on the following 2 variables.
  \describe{
    \item{\code{circ_seq_name}}{Name of circRNA}
    \item{\code{circ_sequence}}{circRNA sequence}
  }
}
\details{
  This is an example sequence data of circRNA.
}
\source{
  This is an example circRNA sequence data.
}
\references{
  R package PCPI.
}
\examples{
  data(circRNA_seq_df)
}
\keyword{datasets}
