\name{protein_seq_df}
\alias{protein_seq_df}
\docType{data}
\title{
  A sample data frame of protein names and sequences.
}
\description{
  This data is a sample data containing the names and sequences of proteins.
}
\usage{data("protein_seq_df")}
\format{
  A data frame with 5 observations on the following 2 variables.
  \describe{
    \item{\code{protein_name}}{Name of protein}
    \item{\code{protein_seq}}{protein sequence}
  }
}
\details{
  This data is an example sequence data of protein.
}
\source{
  This is an example protein sequence data.
}
\references{
  The R package PCPI.
}
\examples{
  data(protein_seq_df)
}
\keyword{datasets}
